requirements.txt
pandas
ipykernel
langchain
python-dotenv
langchain_community
pypdf
bs4
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-openai
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette